What we did for the git commits?
Since w didn't put any descriptions otr titles in to the commits in Github and the only options we left was adding tags in to each commit so we added tags for each changes we did in our website back-end. It's really easy to you to understand by looking add the tags and see what we have changed.